% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{create_observation}
\alias{create_observation}
\title{helper for create_server}
\usage{
create_observation(factor_vars, cont_vars)
}
\arguments{
\item{factor_vars}{vector of strings containing names of factor variables.}

\item{cont_vars}{vector of strings containing names of continous variables.}
}
\description{
returns string defining data.frame for current searched observation.
}
