% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shimex.R
\name{create_shimex}
\alias{create_shimex}
\title{create Shiny App for Exploring Model}
\usage{
create_shimex(explainer, new_obs, data = explainer$data,
  selected_variables = NULL, all = FALSE, main_dir = NULL,
  delete = TRUE)
}
\arguments{
\item{explainer}{explainer created with function `DALEX::explain()`}

\item{new_obs}{a new observation with columns that corresponds to variables used in the model}

\item{data}{---}

\item{selected_variables}{a vector of variable names that will be shown in input panel,}

\item{all}{logical value. If TRUE, then extra tab is displayed showing all explainers,}

\item{main_dir}{string, path where shiny files should be stored.}

\item{delete}{logical value, if TRUE shiny files are deleted on exit of function.}
}
\description{
creates files for Shiny App and runs Shiny.
}
\examples{
\dontrun{
library(randomForest)
model_rm <- randomForest(life_length ~., data = DALEX::dragons, ntree = 200)
explainer <- DALEX::explain(model_rm)
create_shimex(explainer,  DALEX::dragons[1, ])
}

}
