% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny.R
\name{create_shiny}
\alias{create_shiny}
\title{create Shiny App for Exploring Model}
\usage{
create_shiny(explainer, data = NULL, new_obs = NULL,
  selected_variables = NULL, all = F)
}
\arguments{
\item{explainer}{DALEX}

\item{data}{--}

\item{new_obs}{--}

\item{selected_variables}{--}

\item{all}{--}
}
\description{
creates files for Shiny App
}
