% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ui.R
\name{create_ui}
\alias{create_ui}
\title{creates ui.R}
\usage{
create_ui(factor_vars, cont_vars, all)
}
\arguments{
\item{factor_vars}{vector of strings containing names of factor variables.}

\item{cont_vars}{vector of strings containing names of continous variables.}

\item{all}{logical values, if TRUE extra tab is creating showing all explainers together.}
}
\description{
creates string for ui.R file.
}
