% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{.create_server}
\alias{.create_server}
\title{creat server.R}
\usage{
.create_server(factor_vars, cont_vars, all)
}
\arguments{
\item{factor_vars}{vector of strings containing names of factor variables.}

\item{cont_vars}{vector of strings containing names of continous variables.}

\item{all}{logical value. If TRUE, then extra tab is displayed showing all explainers}
}
\description{
creates string for server.R file.
}
