% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ui.R
\name{.create_ui}
\alias{.create_ui}
\title{creates ui.R}
\usage{
.create_ui(selected_explainers)
}
\arguments{
\item{selected_explainers}{list of class of predictors.}
}
\description{
creates string for ui.R file.
}
