% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shimex.R
\name{create_shimex}
\alias{create_shimex}
\title{Create SHIny app for Model EXploration}
\usage{
create_shimex(explainer, new_obs, data = explainer$data,
  selected_variables = NULL, selected_explainers = c("CeterisParibus",
  "BreakDown", "Shap", "LocalModel", "Lime"), main_dir = NULL,
  delete = TRUE, ...)
}
\arguments{
\item{explainer}{explainer created with function `DALEX::explain()`}

\item{new_obs}{a new observation with columns that corresponds to variables used in the model}

\item{data}{---}

\item{selected_variables}{a vector of variable names that will be shown in input panel,}

\item{selected_explainers}{a vector of explainers, that will be shown in the app.
Possible choices: c('CeterisParibus', 'BreakDown', 'Shap', 'Shap_iml', 'Shap_shapper', 'LocalModel', 'Lime', 'Lime_iml')
The default value are explainers implemented by MI2 DataLab.}

\item{main_dir}{string, path where shiny files should be stored.}

\item{delete}{logical value, if TRUE shiny files are deleted on exit of function.}

\item{...}{other parameters.}
}
\description{
creates files for Shiny App and runs Shiny.
}
\examples{
\dontrun{
library(randomForest)
model_rm <- randomForest(life_length ~., data = DALEX::dragons, ntree = 200)
explainer <- DALEX::explain(model_rm)
create_shimex(explainer,  DALEX::dragons[1, ])
}

}
