% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{.create_server}
\alias{.create_server}
\title{creat server.R}
\usage{
.create_server(vars, selected_explainers)
}
\arguments{
\item{vars}{list of class of predictors.}

\item{selected_explainers}{---}
}
\description{
creates string for server.R file.
}
